frontend/pyproject.toml
[project]
name = "frontend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-openai>=0.3.35",
    "mcp-use>=1.3.12",
    "python-dotenv>=1.1.1",
    "streamlit>=1.50.0",
]
